Admin now works
Admin log in at the moment is set for XAMPP, when moved to turnkey password should be included again to allow for actual log in to be checked.

Commented out a few lines, moved the navbar slightly to give more breathing room to the items.

No filtering current is being done on the admin table,
Created a new backup for the database as the old one wasn't loading correctly,
Renamed some key names from first/lastname to just name to reflect the new database set up,
Added a timer to the connection not being successful to inform the user of the outcome.

Fixed a variable name error in the member area to resolve a deleteRequest from not passing,
Still minor errors when the checkbox is not checked.