Update EmailTable.php
Table is now capable of searching via email, can be filtered based on newsletter, deleted, and overall traversed.

Cleaned up some of the page, and also have it ordering in the correct way.

It only requires the 1/0 elements to be converted back into checkboxes or similar to display it in a more user friendly way.